# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * HR Employee Portal API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type EmployeeRole = typeof EmployeeRole[keyof typeof EmployeeRole];


export const EmployeeRole = {
  employee: 'employee',
  admin: 'admin',
} as const;

export interface Employee {
  id: number;
  name: string;
  email: string;
  department: string;
  position: string;
  joinDate: string;
  avatarInitials: string;
  role: EmployeeRole;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  managerId?: number | null;
  /** @nullable */
  managerName?: string | null;
}

export interface LoginInput {
  email: string;
  password: string;
}

export type AttendanceRecordStatus = typeof AttendanceRecordStatus[keyof typeof AttendanceRecordStatus];


export const AttendanceRecordStatus = {
  present: 'present',
  absent: 'absent',
  half_day: 'half_day',
  on_leave: 'on_leave',
  holiday: 'holiday',
  weekend: 'weekend',
} as const;

export interface AttendanceRecord {
  id: number;
  date: string;
  /** @nullable */
  punchIn?: string | null;
  /** @nullable */
  punchOut?: string | null;
  status: AttendanceRecordStatus;
  /** @nullable */
  hoursWorked?: number | null;
  /** @nullable */
  note?: string | null;
}

export type LeaveRequestType = typeof LeaveRequestType[keyof typeof LeaveRequestType];


export const LeaveRequestType = {
  annual: 'annual',
  sick: 'sick',
  casual: 'casual',
  maternity: 'maternity',
  paternity: 'paternity',
  unpaid: 'unpaid',
  other: 'other',
} as const;

export type LeaveRequestStatus = typeof LeaveRequestStatus[keyof typeof LeaveRequestStatus];


export const LeaveRequestStatus = {
  pending: 'pending',
  approved: 'approved',
  rejected: 'rejected',
  cancelled: 'cancelled',
} as const;

export interface LeaveRequest {
  id: number;
  type: LeaveRequestType;
  startDate: string;
  endDate: string;
  days: number;
  status: LeaveRequestStatus;
  reason: string;
  createdAt: string;
  /** @nullable */
  approvedBy?: string | null;
  /** @nullable */
  comments?: string | null;
}

export type LeaveInputType = typeof LeaveInputType[keyof typeof LeaveInputType];


export const LeaveInputType = {
  annual: 'annual',
  sick: 'sick',
  casual: 'casual',
  maternity: 'maternity',
  paternity: 'paternity',
  unpaid: 'unpaid',
  other: 'other',
} as const;

export interface LeaveInput {
  type: LeaveInputType;
  startDate: string;
  endDate: string;
  reason: string;
}

export type LeaveBalanceDetailsItem = {
  type: string;
  total: number;
  used: number;
  remaining: number;
};

export interface LeaveBalance {
  annual: number;
  sick: number;
  casual: number;
  used: number;
  remaining: number;
  details?: LeaveBalanceDetailsItem[];
}

export type HolidayType = typeof HolidayType[keyof typeof HolidayType];


export const HolidayType = {
  public: 'public',
  company: 'company',
  optional: 'optional',
} as const;

export interface Holiday {
  id: number;
  name: string;
  date: string;
  type: HolidayType;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  dayOfWeek?: string | null;
}

export type AnnouncementItemPriority = typeof AnnouncementItemPriority[keyof typeof AnnouncementItemPriority];


export const AnnouncementItemPriority = {
  low: 'low',
  medium: 'medium',
  high: 'high',
} as const;

export type AnnouncementItemType = typeof AnnouncementItemType[keyof typeof AnnouncementItemType];


export const AnnouncementItemType = {
  announcement: 'announcement',
  event: 'event',
} as const;

export interface AnnouncementItem {
  id: number;
  title: string;
  body: string;
  category: string;
  priority: AnnouncementItemPriority;
  type: AnnouncementItemType;
  publishedAt: string;
}

export interface AnnouncementsPage {
  items: AnnouncementItem[];
  total: number;
  page: number;
  limit: number;
}

export type DashboardSummaryMonthlyStats = {
  presentDays: number;
  absentDays: number;
  leaveDays: number;
  totalWorkingDays: number;
  attendanceRate: number;
};

export type DashboardSummaryWeeklyHoursItem = {
  day: string;
  hours: number;
};

export interface DashboardSummary {
  employee: Employee;
  todayAttendance: AttendanceRecord;
  leaveBalance: LeaveBalance;
  upcomingHolidays: Holiday[];
  recentAttendance: AttendanceRecord[];
  monthlyStats: DashboardSummaryMonthlyStats;
  weeklyHours: DashboardSummaryWeeklyHoursItem[];
  recentAnnouncements: AnnouncementItem[];
}

export type ListAttendanceParams = {
/**
 * @nullable
 */
month?: number | null;
/**
 * @nullable
 */
year?: number | null;
};

export type ListHolidaysParams = {
/**
 * @nullable
 */
year?: number | null;
};

export type ListAnnouncementsParams = {
/**
 * @nullable
 */
type?: ListAnnouncementsType;
/**
 * @nullable
 */
page?: number | null;
/**
 * @nullable
 */
limit?: number | null;
};

export type ListAnnouncementsType = typeof ListAnnouncementsType[keyof typeof ListAnnouncementsType] | null;


export const ListAnnouncementsType = {
  announcement: 'announcement',
  event: 'event',
} as const;