# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: HR Employee Portal API
servers:
  - url: /api
    description: Base API path
...
paths:
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Log in with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
        "401":
          description: Invalid credentials
...
  /auth/me:
    get:
      operationId: getCurrentUser
      tags: [auth]
      summary: Get the currently logged-in user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
        "401":
          description: Not authenticated
...
  /leaves:
    get:
      operationId: listLeaves
      tags: [leaves]
      summary: List employee's leave requests
      responses:
        "200":
          description: Leave requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LeaveRequest"
    post:
      operationId: createLeave
      tags: [leaves]
      summary: Apply for a leave
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeaveInput"
      responses:
        "201":
          description: Leave created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeaveRequest"
...
  /leaves/{id}:
    delete:
      operationId: cancelLeave
      tags: [leaves]
      summary: Cancel a leave request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Leave cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeaveRequest"
...
components:
  schemas:
...
    Employee:
      type: object
      required: [id, name, email, department, position, joinDate, avatarInitials, role]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        department:
          type: string
        position:
          type: string
        joinDate:
          type: string
        avatarInitials:
          type: string
        role:
          type: string
          enum: [employee, admin]
        phone:
          type: ["string", "null"]
        managerId:
          type: ["integer", "null"]
        managerName:
          type: ["string", "null"]
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
...
    LeaveRequest:
      type: object
      required: [id, type, startDate, endDate, days, status, reason, createdAt]
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [annual, sick, casual, maternity, paternity, unpaid, other]
        startDate:
          type: string
        endDate:
          type: string
        days:
          type: number
        status:
          type: string
          enum: [pending, approved, rejected, cancelled]
        reason:
          type: string
        createdAt:
          type: string
        approvedBy:
          type: ["string", "null"]
        comments:
          type: ["string", "null"]
    LeaveInput:
      type: object
      required: [type, startDate, endDate, reason]
      properties:
        type:
          type: string
          enum: [annual, sick, casual, maternity, paternity, unpaid, other]
        startDate:
          type: string
        endDate:
          type: string
        reason:
          type: string
...