# lib/api-zod/src/generated/types/employee.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * HR Employee Portal API
 * OpenAPI spec version: 0.1.0
 */
import type { EmployeeRole } from './employeeRole';

export interface Employee {
  id: number;
  name: string;
  email: string;
  department: string;
  position: string;
  joinDate: string;
  avatarInitials: string;
  role: EmployeeRole;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  managerId?: number | null;
  /** @nullable */
  managerName?: string | null;
}